fix(team-builder): 3×2-stack identity row + move level picker into identity
3×2-stack mode now renders identity sprite-LEFT + form-RIGHT instead of
stacking them vertically. At the viewport widths where 3×2-stack fires
(≥2200px → cells ~600–800px) there's enough horizontal room to put the
loadout fields beside the sprite, which keeps each cell shorter and
makes the row easier to scan at a glance. The override is scoped to
[data-layout="3x2-stack"] on the grid so 1×6 / 2×3 / mid-stack modes
keep their vertical-identity layout.

Level picker moved out of the rib and into the identity panel meta row.
Reasons:
- The rib pivots horizontal at full-stack widths (rib becomes a top
  bar), so the vertical level pill no longer had a sensible home there.
- Level naturally belongs with nickname / gender / shiny in the meta
  row above the sprite — it's identity metadata, not a slot decoration.

heroLv is now an interactive popover button (NumberPicker) styled to
match the gender/shiny pills, replacing the previous static <span>.
.ribDecorations now hides entirely at full-stack since the only
remaining child (type pills) doesn't fit in the horizontal rib bar.
